Add (disabled) blueprint support for showing the current beer
Unfortunately accessing GameData via BFL does not seem to work. I found an alternate method to access the current beer via the bar actor.
For now, the beer is added to the top of the same vertical stack layout.
Also, the beer display is disabled pending further testing and refinement. I have to get this code merged then work on other fixes, and merging blueprints seems to be impossible.