Explicitly create HUD and assign specific player owner
The explicit creation replaces the "lazy create", which is a pattern that makes more sense for something like the debug logger.

The "player owner" in the Widget creation step is now assigned specifically to the "local player" (according to FSDGameInstance) rather than being left as default. I've heard that the default player and/or player at index 0 may not always be the local player, whether in multiplayer or returning from missions.