Tweak HUD layout on right side
I changed the origin to the top-right corner instead of right-middle. The intent is to have the top at a very specific spot, and expand down from there. The top should be just below the FPS counter, which is just below the top-right HUD elements. (Before this, the HUD status bar was overlapping with the FPS counter so I had to change it.)

I unchecked "Size to Content" so the preview outline shows up more accurately. It may be better to have this on at some point, but there is no overlapping layout for the moment, and the width is manually matched up with the "rows" anyway.