Add complexity and length icons to the Mission View
For some reason, the Complexity and Duration objects can only be accessed via Mission DNA, and even that must be accessed via function instead of property.
The Mission View widget needed to be widened to help prevent overlap with the mission name, which necessarily widens the HUD and all widgets that go into it.